指摘部分を修正
