+++ SuperNewRoles/Roles/CrewMate/DyingMessenger.cs
@@ -1,14 +1,12 @@
 using System;
 using System.Collections.Generic;
-using HarmonyLib;
-using SuperNewRoles.CustomCosmetics;
 using UnityEngine;
 
 namespace SuperNewRoles.Roles.Crewmate;
 
 public class DyingMessenger
 {
-    private const int OptionId = 1171;
+    private const int OptionId = 1179;
     public static CustomRoleOption DyingMessengerOption;
     public static CustomOption DyingMessengerPlayerCount;
     public static CustomOption DyingMessengerGetRoleTime;
@@ -23,42 +21,10 @@ public class DyingMessenger
 
     public static List<PlayerControl> DyingMessengerPlayer;
     public static Color32 color = new(191, 197, 202, byte.MaxValue);
-    public static Dictionary<byte, (DateTime, PlayerControl)> KillTime;
+    public static Dictionary<byte, (DateTime, PlayerControl)> ActualDeathTime;
     public static void ClearAndReload()
     {
         DyingMessengerPlayer = new();
-        KillTime = new();
-    }
-
-    [HarmonyPatch(typeof(MeetingHud), nameof(MeetingHud.Update))]
-    public class MeetingHudUpdatePatch
-    {
-        public static void Postfix()
-        {
-            if (MeetingHud.Instance)
-            {
-                foreach (PlayerVoteArea player in MeetingHud.Instance.playerStates)
-                {
-                    PlayerControl target = null;
-                    PlayerControl.AllPlayerControls.ToList().ForEach(x =>
-                    {
-                        string name = player.NameText.text.Replace(GetLightAndDarkerText(true), "").Replace(GetLightAndDarkerText(false), "");
-                        if (name == x.Data.PlayerName) target = x;
-                    });
-                    if (target != null)
-                    {
-                        if (ConfigRoles.IsLightAndDarker.Value)
-                        {
-                            if (player.NameText.text.Contains(GetLightAndDarkerText(true)) ||
-                                player.NameText.text.Contains(GetLightAndDarkerText(false))) continue;
-                            player.NameText.text += GetLightAndDarkerText(CustomColors.lighterColors.Contains(target.Data.DefaultOutfit.ColorId));
-                        }
-                        else player.NameText.text = player.NameText.text.Replace(GetLightAndDarkerText(true), "").Replace(GetLightAndDarkerText(false), "");
-                    }
-                    else Logger.Error($"プレイヤーコントロールを取得できませんでした。 プレイヤー名 : {player.NameText.text}", "LightAndDarkerText");
-                }
-            }
-        }
-        public static string GetLightAndDarkerText(bool isLight) => $" ({(isLight ? ModTranslation.GetString("LightColor") : ModTranslation.GetString("DarkerColor"))[0]})";
+        ActualDeathTime = new();
     }
 }
